Move Cannon subsystem to subsystems package.
Might cause merging problems, as Simon and I both were making changes around
this cannon subsystem. You have been warned.
