+++ src/org/robockets/tshirtcannon/RobotMap.java
@@ -1,5 +1,7 @@
 package org.robockets.tshirtcannon;
 
+import org.robockets.tshirtcannon.subsystems.Cannon;
+
 /**
  * The RobotMap is a mapping from the ports sensors and actuators are wired into
  * to a variable name. This provides flexibility changing wiring, makes checking
